src/requirements.txt
Flask==3.*
Flask-SocketIO==5.*
azure-cosmos==4.*
azure-identity==1.*
eventlet==0.*
gunicorn==21.*